% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcGWAS-package.R
\docType{package}
\name{cbcGWAS-package}
\alias{cbcGWAS}
\alias{cbcGWAS-package}
\title{Liability-scale conversion for covariate-adjusted binary GWAS}
\description{
Conversion factors, calibration, bias-estimation interfaces, and simulations
under a Gaussian liability model. Start with \code{\link[=lambda_working]{lambda_working()}} or
\code{\link[=run_simulations]{run_simulations()}}.
}
\author{
\strong{Maintainer}: Siyi Chen \email{sche11@lsuhsc.edu}

Authors:
\itemize{
  \item Siyi Chen \email{sche11@lsuhsc.edu}
}

}
\keyword{internal}
